+++ contracts/multisig-improved/src/action_types/mod.rs
@@ -1,4 +1,5 @@
 pub mod discard;
+pub mod execute_action;
 pub mod perform;
 pub mod perform_endpoints;
 pub mod propose;
