+++ contracts/multisig-improved/src/external/views.rs
@@ -11,6 +11,7 @@ pub trait ViewsModule:
     + crate::action_types::propose::ProposeModule
     + crate::action_types::sign::SignModule
     + crate::action_types::perform::PerformModule
+    + crate::action_types::execute_action::ExecuteActionModule
     + crate::check_signature::CheckSignatureModule
     + super::events::EventsModule
 {
