+++ contracts/multisig-improved/src/multisig_improved.rs
@@ -17,6 +17,7 @@ pub trait Multisig:
     + check_signature::CheckSignatureModule
     + action_types::propose_endpoints::ProposeEndpointsModule
     + action_types::perform_endpoints::PerformEndpointsModule
+    + action_types::execute_action::ExecuteActionModule
     + action_types::propose::ProposeModule
     + action_types::sign::SignModule
     + action_types::perform::PerformModule
